Rename data.rs to group_data.rs
